requirements.txt
# LLM Abstraction & Observability
litellm>=1.40.0
langfuse>=2.0.0

# Workflow & AI
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
pyyaml>=6.0.0

# MCP
mcp>=1.0.0

# PDF Processing
pypdf2>=3.0.0
pdfplumber>=0.11.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# UI
streamlit>=1.30.0
plotly>=5.18.0

# API Server
flask>=3.0.0
flask-cors>=4.0.0

# Development
black>=24.0.0
mypy>=1.8.0
ruff>=0.1.0

# Optional (for future scale)
# redis>=5.0.0
# aioredis>=2.0.0
